Update landing page module summaries
