+++ bims/static/css/landing_page.css
@@ -223,8 +223,11 @@ header.masthead p {
 
 .chart-icon-wrapper {
     position: relative;
-    width: 200px;
-    height: 200px;
+    display: flex;
+    width: 100%;
+    height: 150px;
+    justify-content: center;
+    align-items: center;
     margin: 0 auto;
 }
 
@@ -354,9 +357,6 @@ html, body {
 
 .donut-chart {
     position: absolute !important;
-    top: 10px;
-    left: 50%;
-    transform: translate(-50%, 0);
 }
 
 .btn-account {
@@ -450,9 +450,15 @@ html, body {
 }
 
 .summary-data {
-    margin-top: 50px;
+    margin-top: 30px;
     text-align: center;
     color: #263c4c;
+    display: flex;
+    justify-content: center;
+}
+
+.module-numbers {
+    font-weight: bold;
 }
 
 .summary-data .fa {
@@ -462,7 +468,7 @@ html, body {
 .summary-data p {
     margin-top: 20px;
     font-size: 20pt;
-    font-weight: bold;
+    color: var(--main-accent-color);
 }
 
 .biodiversity-data-container {
@@ -547,11 +553,10 @@ html, body {
     margin-left: 10px;
     width: 50px
 }
+
 .module-icon {
+    width: 70px !important;
     position: absolute;
-    margin-left: -50px;
-    width: 100px;
-    z-index: 0;
 }
 
 .module-name {
