+++ bims/static/js/landing-page.js
@@ -55,3 +55,124 @@ document.addEventListener('DOMContentLoaded', function (event) {
         generateDoughnutChart(chartContainer, config);
     });
 });
+
+function createSummaryChart(data){
+    if(!data){
+        return;
+    }
+    let bColor;
+    let labels = [];
+    let dataChart = [];
+
+
+    if(data.hasOwnProperty('conservation-status')){
+        let lc = 0;
+        let nt = 0;
+        let vu = 0;
+        let en = 0;
+        let dd = 0;
+        let ne = 0;
+        let ce = 0;
+
+        let anuraData = data['conservation-status'];
+        if (anuraData.hasOwnProperty('Least concern')) {
+            lc += anuraData['Least concern'];
+        }
+        if (anuraData.hasOwnProperty('Near threatened')) {
+            nt += anuraData['Near threatened'];
+        }
+        if (anuraData.hasOwnProperty('Vulnerable')) {
+            vu += anuraData['Vulnerable'];
+        }
+        if (anuraData.hasOwnProperty('Endangered')) {
+            en += anuraData['Endangered'];
+        }
+        if (anuraData.hasOwnProperty('Not evaluated')) {
+            ne += anuraData['Not evaluated'];
+        }
+        if (anuraData.hasOwnProperty('Critically endangered')) {
+            ce += anuraData['Critically endangered'];
+        }
+        if (anuraData.hasOwnProperty('Data Deficient')) {
+            dd += anuraData['Data Deficient'];
+        }
+
+        labels = ["Data Deficient", "Not evaluated", "Least concern", "Near threatened", "Vulnerable", "Endangered", "Critically endangered"],
+            dataChart = [dd, ne, lc, nt, vu, en, ce];
+        bColor = [
+            '#D7CD47',
+            '#39B2A3',
+            '#17766B',
+            '#2C495A',
+            '#525351',
+            '#8D2641',
+            '#641f30',
+        ];
+    }
+    else if(data.hasOwnProperty('endemism')){
+
+        for (let key in data['endemism']) {
+            dataChart.push(data['endemism'][key]);
+            labels.push(key)
+        }
+        bColor = [
+            '#a13447',
+            '#00a99d',
+            '#e0d43f'
+        ];
+    }
+
+    else if (data.hasOwnProperty('division')) {
+        $.each(data['division'], function (index, divisionData) {
+            if (divisionData['taxonomy__additional_data__Division'] == null) {
+                labels.push('Unknown');
+            } else {
+                labels.push(divisionData['taxonomy__additional_data__Division']);
+            }
+            dataChart.push(divisionData['count']);
+        });
+        bColor = [
+            '#8D2641', '#641f30',
+            '#E6E188', '#D7CD47',
+            '#9D9739', '#525351',
+            '#618295', '#2C495A',
+            '#39B2A3', '#17766B',
+            '#859FAC', '#1E2F38'
+        ];
+
+
+    }
+    else if (data.hasOwnProperty('origin')) {
+
+        for (let key in data['origin']) {
+            dataChart.push(data['origin'][key]);
+            labels.push(key)
+        }
+        bColor = [
+            '#a13447',
+            '#e0d43f'
+        ];
+    }else {
+        $.each(data['ecological_data'], function (index, ecologicalData) {
+            labels.push(ecologicalData['value']);
+            dataChart.push(ecologicalData['count']);
+        });
+        bColor = [
+            '#641f30',
+            '#8D2641',
+            '#D7CD47',
+            '#17766B',
+            '#39B2A3'
+        ];
+    }
+
+    return {
+        labels: labels,
+        datasets: [{
+            data: dataChart,
+            backgroundColor: bColor,
+            borderWidth: 1
+        }]
+    };
+}
+
